fix: unresolved reference on prod builds
