+++ app/src/main/kotlin/com/wire/android/navigation/NavigationUtils.kt
@@ -20,16 +20,9 @@
 
 package com.wire.android.navigation
 
-import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.runtime.Composable
 import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavController
-import androidx.navigation.NavDestination
-import com.datadog.android.compose.ExperimentalTrackingApi
-import com.datadog.android.compose.NavigationViewTrackingEffect
-import com.datadog.android.rum.tracking.AcceptAllNavDestinations
-import com.google.accompanist.navigation.animation.rememberAnimatedNavController
 import com.wire.android.appLogger
 
 @ExperimentalMaterial3Api
@@ -97,16 +90,4 @@ fun String.getPrimaryRoute(): String {
         else -> this
     }
     return primaryRoute
-}
-
-@OptIn(ExperimentalAnimationApi::class, ExperimentalTrackingApi::class)
-@Composable
-fun rememberTrackingAnimatedNavController(nameFromRoute: (String) -> String?) = rememberAnimatedNavController().apply {
-    NavigationViewTrackingEffect(
-        navController = this,
-        trackArguments = true,
-        destinationPredicate = object : AcceptAllNavDestinations() {
-            override fun getViewName(component: NavDestination): String? = component.route?.let { nameFromRoute(it) }
-        }
-    )
 }
