+++ app/src/prod/kotlin/com/wire/android/navigation/TrackingNavControler.kt
@@ -1,0 +1,9 @@
+package com.wire.android.navigation
+
+import androidx.compose.animation.ExperimentalAnimationApi
+import androidx.compose.runtime.Composable
+import com.google.accompanist.navigation.animation.rememberAnimatedNavController
+
+@OptIn(ExperimentalAnimationApi::class)
+@Composable
+fun rememberTrackingAnimatedNavController(nameFromRoute: (String) -> String?) = rememberAnimatedNavController()
